# cloudnet-rest-module/src/main/resources/documentation/swagger.yaml
openapi: 3.1.0
info:
  description: |
    In order to access rest API v3 you will have to authenticate yourself using basic auth.
    To create a basic auth user use the `rest user create <id> <password>` command on your cloudnet node.
    After authenticating with basic auth on `/auth` you will get a bearer jwt access token that must be used for further
    requests to the api. For further information about the response take a look at the `/auth` endpoint.
  version: '3'
  title: CloudNet
  contact:
    name: CloudNet
    url: https://cloudnetservice.eu
  license:
    name: Apache license, version 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: '{protocol}://{host}:{port}/api/v3/'
    variables:
      protocol:
        enum:
          - http
          - https
        default: https
      host:
        default: 127.0.0.1
      port:
        default: '2812'
security:
  - bearerAuth: [ ]
...
paths:
...
  /cluster:
    get:
      tags:
        - Cluster
      summary: Lists all nodes
      description: |
        Lists all nodes which are registered in the node. This includes nodes
        which are currently not connected to the node. If they are not connected
        available will be `false` and the `nodeInfoSnapshot` will be null.

        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:cluster_read`
          - `cloudnet_rest:cluster_node_list`
      operationId: cluster
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  nodes:
                    type: array
                    items:
                      $ref: '#/components/schemas/NodeServer'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
    put:
      requestBody:
        required: true
        description: The new node configuration entry.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkClusterNode'
      tags:
        - Cluster
      summary: Updates a node
      description: |
        Changes the configuration of a specific node in the local node's config
        file. The update will be pushed live to the node cluster settings. That
        does not mean that the node tries to connect to the updated node.

        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:cluster_write` 
          - `cloudnet_rest:cluster_node_update`
      operationId: clusterUpdate
      responses:
        '204':
          description: The node was updated successfully
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
    post:
      requestBody:
        required: true
        description: The node configuration entry.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkClusterNode'
      tags:
        - Cluster
      summary: Creates a new node
      description: |
        Creates a node in the node's local configuration and updates it live to
        the node cluster settings. That does not mean that the node tries to
        connect to the newly registered node.

        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:cluster_write`
          - `cloudnet_rest:cluster_node_create`
      operationId: clusterCreate
      responses:
        '201':
          description: The new node was created successfully
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
...
  /node:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Node
      summary: Node info
      description: |
        Information about the node the rest listener belongs to.

        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:node_read`
          - `cloudnet_rest:node_info`
      operationId: generalInfo
      responses:
        '200':
          description: Information about the node the request was sent to
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeInfo'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
...
  /node/logLines:
    get:
      tags:
        - Node
      summary: Get the cached log lines of the node the request is sent to
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:node_read`
          - `cloudnet_rest:node_log_lines`
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  lines:
                    type: array
                    items:
                      type: string
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /node/reload:
    parameters:
      - name: type
        in: query
        required: false
        description: The reload type
        schema:
          type: string
          enum:
            - all
            - config
          default: all
    post:
      tags:
        - Node
      summary: Reloads the node the request is sent to
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:node_write`
          - `cloudnet_rest:node_reload`
      responses:
        '204':
          description: "The node was reloaded."
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /node/config:
    get:
      tags:
        - Node
      summary: Get the current configuration file of the node the request is sent to
      description: |
        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:node_read`
          - `cloudnet_rest:node_config_get`
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeConfiguration'
          description: The node configuration that was requested
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeConfiguration'
      tags:
        - Node
      summary: Updates the configuration file of the node the request is sent to
      description: |
        To add/remove a node from the configuration please use the cluster 
        routes to add/remove a node.

        One of the following scopes is needed to execute the request:
          - `cloudnet_rest:node_write`
          - `cloudnet_rest:node_config_update`
      responses:
        '204':
          description: The node configuration was updated.
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  responses:
    Problem:
      description: The action couldn't be executed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      description: |
        An object describing a problem as specific in RFC 9457. Note that this
        object can contain any number of additional properties depending on the
        called route and error cause.
      type: object
      properties:
        type:
          type: string
          format: uri
          description: An absolute URI that identifies the problem type
          default: about:blank
        title:
          type: string
          description: |
            A short summary of the problem type. Written in English and readable
            for engineers (usually not suited for non technical stakeholders and
            not localized).
          examples:
            - Service Unavailable
        status:
          type: integer
          format: int32
          description: |
            The HTTP status code generated by the origin server for this
            occurrence of the problem.
          minimum: 400
          exclusiveMaximum: 600
          examples:
            - 503
        detail:
          type: string
          description: |
            A human-readable explanation specific to this occurrence of the
            problem
        instance:
          type: string
          format: uri
          description: |
            An absolute URI reference that identifies the specific occurrence of
            the problem. It may or may not yield further information if
            dereferenced.
...
    NetworkClusterNode:
      type: object
      allOf:
        - $ref: '#/components/schemas/JsonDocPropertyable'
        - properties:
            uniqueId:
              type: string
              examples:
                - Node-1
            listeners:
              type: array
              items:
                $ref: '#/components/schemas/HostAndPort'
    NodeConfiguration:
      type: object
      allOf:
        - $ref: '#/components/schemas/JsonDocPropertyable'
        - properties:
            identity:
              $ref: '#/components/schemas/NetworkClusterNode'
            clusterConfig:
              type: object
              properties:
                clusterId:
                  type: string
                  examples:
                    - a0e8765e-7853-455b-bb2b-7a022a607be2
                nodes:
                  type: array
                  items:
                    $ref: '#/components/schemas/NetworkClusterNode'
            ipWhitelist:
              type: array
              items:
                type: string
              examples:
                - 192.168.56.1
                - 127.0.0.1
            maxCPUUsageToStartServices:
              type: number
              examples:
                - 90
            maxMemory:
              type: number
              examples:
                - 2048
            maxServiceConsoleLogCacheSize:
              type: integer
              examples:
                - 128
            processTerminationTimeoutSeconds:
              type: integer
              examples:
                - 5
            forceInitialClusterDataSync:
              type: boolean
              examples:
                - true
            printErrorStreamLinesFromServices:
              type: boolean
              examples:
                - true
            runBlockedServiceStartTryLaterAutomatic:
              type: boolean
              examples:
                - true
            jvmCommand:
              type: string
              examples:
                - java
            hostAddress:
              type: string
              examples:
                - 127.0.0.1
            connectHostAddress:
              type: string
              examples:
                - 127.0.0.1
            httpListeners:
              type: array
              items:
                $ref: '#/components/schemas/HostAndPort'
            clientSslConfig:
              $ref: '#/components/schemas/ConfigurationOptionSSL'
            serverSslConfig:
              $ref: '#/components/schemas/ConfigurationOptionSSL'
            webSslConfig:
              $ref: '#/components/schemas/ConfigurationOptionSSL'
...
    ConfigurationOptionSSL:
      type: object
      properties:
        enabled:
          type: boolean
          examples:
            - false
        clientAuth:
          type: boolean
          examples:
            - true
        trustCertificatePath:
          type: string
          examples:
            - null
        certificatePath:
          type: string
          examples:
            - local/certificate.pem
        privateKeyPath:
          type: string
          examples:
            - local/privateKey.key
    NodeServer:
      type: object
      properties:
        node:
          $ref: '#/components/schemas/NetworkClusterNode'
        state:
          type: string
          enum:
            - UNAVAILABLE
            - SYNCING
            - READY
            - DISCONNECTED
        head:
          type: boolean
          description: If the node server is the current head node.
        local:
          type: boolean
          description: If the node server is the node that handled the request.
        nodeInfoSnapshot:
          $ref: '#/components/schemas/NodeInfoSnapshot'
    NetworkClusterNodeInfoSnapshot:
      type: object
      allOf:
        - $ref: '#/components/schemas/JsonDocPropertyable'
        - properties:
            node:
              $ref: '#/components/schemas/NetworkClusterNode'
            state:
              type: string
              enum:
                - UNAVAILABLE
                - SYNCING
                - READY
                - DISCONNECTED
              examples:
                - READY
            head:
              type: boolean
              examples:
                - false
            local:
              type: boolean
              examples:
                - true
            nodeInfoSnapshot:
              $ref: '#/components/schemas/NodeInfoSnapshot'
    NodeInfoSnapshot:
      type: object
      allOf:
        - $ref: '#/components/schemas/JsonDocPropertyable'
        - properties:
            creationTime:
              type: integer
              description: The unix epoch millis when the snapshot was created
              examples:
                - 1620819600457
            startupMillis:
              type: integer
              description: The unix epoch millis when the node was started
              examples:
                - 1620819600457
            maxMemory:
              type: number
              description: The maximum memory amount (in MB) the node can consume
              examples:
                - 2048
            usedMemory:
              type: integer
              description: The memory (in MB) that is used by services on the node
              examples:
                - 512
            reservedMemory:
              type: integer
              description: The memory (in MB) that is reserved for services on the node
              examples:
                - 725
            currentServicesCount:
              type: integer
              description: The amount of services that are registered on the node
              examples:
                - 2
            drain:
              type: boolean
              description: Indicates if the node is draining (stopping after all running services were stopped)
              examples:
                - true
            node:
              $ref: '#/components/schemas/NetworkClusterNode'
            version:
              $ref: '#/components/schemas/CloudNetVersion'
            processSnapshot:
              $ref: '#/components/schemas/ProcessSnapshot'
            maxCPUUsageToStartServices:
              type: number
              examples:
                - 90.0
            modules:
              type: array
              items:
                $ref: '#/components/schemas/ModuleConfiguration'
...
    JsonDocPropertyable:
      type: object
      properties:
        properties:
          type: object
          examples:
            - { }
...
    NodeInfo:
      type: object
      properties:
        version:
          $ref: '#/components/schemas/CloudNetVersion'
        nodeInfoSnapshot:
          $ref: '#/components/schemas/NetworkClusterNodeInfoSnapshot'
        lastNodeInfoSnapshot:
          $ref: '#/components/schemas/NetworkClusterNodeInfoSnapshot'
        serviceCount:
          type: integer
          examples:
            - 0
        clientConnections:
          type: array
          items:
            type: string
            examples:
              - 127.0.0.1:1411
              - 127.0.0.1:1412
    ProcessSnapshot:
      type: object
      properties:
        pid:
          type: integer
          examples:
            - 8168
        cpuUsage:
          type: number
          examples:
            - 0.7781862806132257
        systemCpuUsage:
          type: number
          examples:
            - 0.7781862806132257
        maxHeapMemory:
          type: number
          examples:
            - 2111832064
        heapUsageMemory:
          type: number
          examples:
            - 230091112
        noHeapUsageMemory:
          type: number
          examples:
            - 46906136
        unloadedClassCount:
          type: number
          examples:
            - 21
        totalLoadedClassCount:
          type: number
          examples:
            - 5940
        currentLoadedClassCount:
          type: number
          examples:
            - 5919
        threads:
          type: array
          items:
            $ref: '#/components/schemas/ThreadSnapshot'
    ThreadSnapshot:
      type: object
      properties:
        id:
          type: number
          examples:
            - 5
        priority:
          type: integer
          examples:
            - 1
        daemon:
          type: boolean
          examples:
            - true
        name:
          type: string
          examples:
            - WindowsStreamPump
        threadState:
          type: string
          enum:
            - NEW
            - RUNNABLE
            - BLOCKED
            - WAITING
            - TIMED_WAITING
            - TERMINATED
          examples:
            - RUNNABLE
    ModuleConfiguration:
      type: object
      properties:
        runtimeModule:
          type: boolean
          examples:
            - false
        storesSensitiveData:
          type: boolean
          examples:
            - true
        group:
          type: string
          examples:
            - eu.cloudnetservice.modules
        name:
          type: string
          examples:
            - CloudNet-CloudFlare
        version:
          type: string
          examples:
            - '1.1'
        main:
          type: string
          examples:
            - eu.cloudnetservice.modules.cloudflare.CloudNetCloudflareModule
        description:
          type: string
          examples:
            - Node extension for the https://cloudflare.com REST-API
        author:
          type: string
          examples:
            - CloudNetService
        website:
          type: string
          examples:
            - https://cloudnetservice.eu
        dataFolder:
          type: string
          examples:
            - null
        repositories:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                examples:
                  - CloudNet
              url:
                type: string
                examples:
                  - https://repo.cloudnetservice.eu/
        dependencies:
          type: array
          items:
            type: object
            properties:
              repo:
                type: string
                examples:
                  - CloudNet
              url:
                type: string
                examples:
                  - https://repo.cloudnetservice.eu/
              group:
                type: string
                examples:
                  - eu.cloudnetservice.cloudnet
              name:
                type: string
                examples:
                  - CloudNet-Bridge
              version:
                type: string
                examples:
                  - '1.1'
        minJavaVersionId:
          type: integer
          examples:
            - 11
        properties:
          type: object
          examples:
            - { }
...
    HostAndPort:
      type: object
      properties:
        host:
          type: string
          examples:
            - 127.0.0.1
        port:
          type: integer
          examples:
            - 1410
...
    CloudNetVersion:
      type: object
      properties:
        major:
          type: integer
          examples:
            - 4
        minor:
          type: integer
          examples:
            - 0
        patch:
          type: integer
          examples:
            - 1
        revision:
          type: string
          examples:
            - 5dd7f63
        versionType:
          type: string
          examples:
            - RELEASE
        versionTitle:
          type: string
          examples:
            - Blizzard